Remove runit dependency
The previous commits introduce unnecessary dependencies since
they only monitor process state. This commit removes the runit
dependency in favor of the docker native restart policy option
of "always".

As a result of the switch to the docker cookbook, it now explicitly
sets environment variables on the container instead of using the
env-file option.