Pivot to app cookbook design
This commit pivots away from the original idea since a good drone
cookbook already exists. It changes the focus of this cookbook towards
deploying the drone application on a given node with very little
configurability as though it was a bundled app of sorts. It also
provides a front facing nginx ssl proxy with certificates managed via
letsencrypt as the original cookbook wanted to do.